+++ src/main/java/insane96mcp/enhancedai/modules/base/BaseModule.java
@@ -1,9 +1,9 @@
 package insane96mcp.enhancedai.modules.base;
 
 
-import insane96mcp.enhancedai.modules.base.feature.AvoidExplosionsFeature;
-import insane96mcp.enhancedai.modules.base.feature.BaseFeature;
-import insane96mcp.enhancedai.modules.base.feature.TargetingFeature;
+import insane96mcp.enhancedai.modules.base.feature.AvoidExplosions;
+import insane96mcp.enhancedai.modules.base.feature.Base;
+import insane96mcp.enhancedai.modules.base.feature.Targeting;
 import insane96mcp.enhancedai.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -11,16 +11,16 @@ import insane96mcp.insanelib.base.Module;
 @Label(name = "Base")
 public class BaseModule extends Module {
 
-	public BaseFeature base;
-	public TargetingFeature targeting;
-	public AvoidExplosionsFeature avoidExplosions;
+	public Base base;
+	public Targeting targeting;
+	public AvoidExplosions avoidExplosions;
 
 	public BaseModule() {
 		super(Config.builder);
 		this.pushConfig(Config.builder);
-		base = new BaseFeature(this);
-		targeting = new TargetingFeature(this);
-		avoidExplosions = new AvoidExplosionsFeature(this);
+		base = new Base(this);
+		targeting = new Targeting(this);
+		avoidExplosions = new AvoidExplosions(this);
 		Config.builder.pop();
 	}