+++ src/main/java/insane96mcp/enhancedai/modules/creeper/CreeperModule.java
@@ -1,6 +1,6 @@
 package insane96mcp.enhancedai.modules.creeper;
 
-import insane96mcp.enhancedai.modules.creeper.feature.CreeperSwellFeature;
+import insane96mcp.enhancedai.modules.creeper.feature.CreeperSwell;
 import insane96mcp.enhancedai.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -8,12 +8,12 @@ import insane96mcp.insanelib.base.Module;
 @Label(name = "Creeper")
 public class CreeperModule extends Module {
 
-	CreeperSwellFeature creeperSwell;
+	CreeperSwell creeperSwell;
 
 	public CreeperModule() {
 		super(Config.builder);
 		this.pushConfig(Config.builder);
-		creeperSwell = new CreeperSwellFeature(this);
+		creeperSwell = new CreeperSwell(this);
 		Config.builder.pop();
 	}