+++ src/main/java/insane96mcp/enhancedai/modules/creeper/feature/CreeperSwell.java
@@ -24,7 +24,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import java.util.ArrayList;
 
 @Label(name = "Creeper Swell", description = "Various changes to Creepers exploding. Ignoring Walls, Walking Fuse and smarter exploding based off explosion size")
-public class CreeperSwellFeature extends Feature {
+public class CreeperSwell extends Feature {
 
 	private final ForgeConfigSpec.ConfigValue<Double> cenaChanceConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> walkingFuseConfig;
@@ -40,7 +40,7 @@ public class CreeperSwellFeature extends Feature {
 	public double launch = 0.05;
 	public boolean tntLike = false;
 
-	public CreeperSwellFeature(Module module) {
+	public CreeperSwell(Module module) {
 		super(Config.builder, module);
 		Config.builder.comment(this.getDescription()).push(this.getName());
 		cenaChanceConfig = Config.builder