+++ src/main/java/insane96mcp/enhancedai/modules/skeleton/SkeletonModule.java
@@ -1,6 +1,6 @@
 package insane96mcp.enhancedai.modules.skeleton;
 
-import insane96mcp.enhancedai.modules.skeleton.feature.SkeletonAIFeature;
+import insane96mcp.enhancedai.modules.skeleton.feature.SkeletonAI;
 import insane96mcp.enhancedai.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -8,12 +8,12 @@ import insane96mcp.insanelib.base.Module;
 @Label(name = "Skeleton")
 public class SkeletonModule extends Module {
 
-	public SkeletonAIFeature skeletonAI;
+	public SkeletonAI skeletonAI;
 
 	public SkeletonModule() {
 		super(Config.builder);
 		this.pushConfig(Config.builder);
-		skeletonAI = new SkeletonAIFeature(this);
+		skeletonAI = new SkeletonAI(this);
 		Config.builder.pop();
 	}