+++ src/main/java/insane96mcp/enhancedai/modules/spider/SpiderModule.java
@@ -1,7 +1,7 @@
 package insane96mcp.enhancedai.modules.spider;
 
-import insane96mcp.enhancedai.modules.spider.feature.MiscFeature;
-import insane96mcp.enhancedai.modules.spider.feature.ThrowingWebFeature;
+import insane96mcp.enhancedai.modules.spider.feature.Misc;
+import insane96mcp.enhancedai.modules.spider.feature.ThrowingWeb;
 import insane96mcp.enhancedai.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -9,14 +9,14 @@ import insane96mcp.insanelib.base.Module;
 @Label(name = "Spider")
 public class SpiderModule extends Module {
 
-	public ThrowingWebFeature throwingWeb;
-	public MiscFeature misc;
+	public ThrowingWeb throwingWeb;
+	public Misc misc;
 
 	public SpiderModule() {
 		super(Config.builder);
 		this.pushConfig(Config.builder);
-		this.throwingWeb = new ThrowingWebFeature(this);
-		this.misc = new MiscFeature(this);
+		this.throwingWeb = new ThrowingWeb(this);
+		this.misc = new Misc(this);
 		Config.builder.pop();
 	}